[#90] feat: 수익률 자동 계산 기능 구현

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/basic/dto/RealEstateListDto.java
@@ -12,7 +12,7 @@ public class RealEstateListDto extends RealEstateDto {
 
     public String managerName;
     public int salePrice;
-    public int revenueRate;
+    public double revenueRate;
     public double platArea;
     public double totArea;
     public double archArea;
